Moved validator constraints from core to bundle
