+++ src/PrestaShopBundle/Validator/Constraint/IsUrlRewrite.php
@@ -24,24 +24,17 @@
  * International Registered Trademark & Property of PrestaShop SA
  */
 
-namespace PrestaShop\PrestaShop\Core\ConstraintValidator\Constraints;
+namespace PrestaShop\PrestaShop\PrestaShopBundle\Validator\Constraint;
 
-use PrestaShop\PrestaShop\Core\ConstraintValidator\DefaultLanguageValidator;
 use Symfony\Component\Validator\Constraint;
 
 /**
- * Class DefaultLanguage is responsible for checking if the array contains default language id - its common to require
- * default language to be presented when saving required multi-language fields.
+ * Class IsUrlRewrite is responsible of validating url rewrites according to several patterns
+ * which differ when ascending urls are enabled or not.
+ *
+ * @Annotation
  */
-class DefaultLanguage extends Constraint
+class IsUrlRewrite extends Constraint
 {
-    public $message = 'The field %field_name% is required at least in your default language.';
-
-    /**
-     * {@inheritdoc}
-     */
-    public function validatedBy()
-    {
-        return DefaultLanguageValidator::class;
-    }
+    public $message = '%s is invalid.';
 }
